+++ packages/lib/src/transforms/bigint.ts
@@ -1,0 +1,16 @@
+import type { ModelPropTransform } from "../modelShared/prop"
+
+const _stringToBigIntTransform: ModelPropTransform<string, bigint> = {
+  transform({ originalValue, cachedTransformedValue }) {
+    return cachedTransformedValue ?? BigInt(originalValue)
+  },
+
+  untransform({ transformedValue, cacheTransformedValue }) {
+    if (typeof transformedValue === "bigint") {
+      cacheTransformedValue()
+    }
+    return transformedValue.toString()
+  },
+}
+
+export const stringToBigIntTransform = () => _stringToBigIntTransform